requirements.txt
# Core dependencies
pymupdf==1.24.10
streamlit==1.38.0
python-pptx==1.0.2
Pillow==10.4.0
requests==2.32.3
frontend==0.0.3

# LLM and AI dependencies
llama-index-core==0.10.58
llama-index-readers-file==0.1.30
llama-index-llms-nvidia==0.1.4
llama-index-embeddings-nvidia==0.1.4
llama-index-vector-stores-milvus==0.1.20
langchain>=0.0.267

# Database dependencies
pymilvus==2.4.4
pymongo>=4.5.0

# Web framework dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6

# Frontend dependencies
jinja2>=3.1.2
bootstrap5>=0.1.5

# Utility dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0